refactor bottom header
