+++ components/Article/ArticleNavigation.module.css
@@ -1,0 +1,3 @@
+.tab {
+    
+}
