+++ pages/home.module.css
@@ -2,58 +2,13 @@
 
 }
 
-/*=============== Header ===============*/
+/*=============== TopHeader ===============*/
 .header {
     position: fixed;
     top: 0;
     left: 0;
     width: 100%;
     z-index: 1;
-}
-
-.header__top {
-    height: 60px;
-    /*border-bottom: 1px solid #f1f1f1;*/
-}
-
-.header__container {
-    margin: 0 auto;
-    width: 1440px;
-    height: 100%;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-}
-
-/*=============== Navigation ===============*/
-.header__navigation {
-    height: 100%;
-    display: flex;
-    flex-direction: row;
-}
-
-.header__navlist {
-    display: flex;
-    flex-direction: row;
-    height: 100%;
-}
-
-.header__navlink {
-    display: flex;
-    align-items: center;
-    height: 100%;
-    padding: 0 14px;
-    text-decoration: none;
-    font-size: 14px;
-    font-weight: 500;
-    transition: all 0.3s ease;
-}
-
-.header__banner {
-    height: 100%;
-    width: 139px;
-    display: flex;
-    align-items: center;
 }
 
 
@@ -99,9 +54,6 @@
 .main__right {
 }
 
-.header__rightside {
-
-}
 
 .author__info,
 .related__articles,
